chore: create docker files for build and deploy
Changes:

travis-ci now only needs docker.
create a docker file with scala and sbt (docker-ci/Dockerfile) this will be the base for all other Dockerfiles.
create a docker file that extends the base one and has responsability of running the tests (docker-ci/Dockerfile.build).
create a docker file that only responsability is to run the jar.
create a script for running postgresql in case the dev wants to run the service locally (scripts/start-forumdb.sh).
create a docker-compose that can be run running `sh /scripts/start.sh`.